[catnap] Enhancement: Add socket files
